% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{celcius}
\alias{celcius}
\title{Fahrenheit to Celsius}
\usage{
celcius(tempF)
}
\arguments{
\item{tempC}{temperature value(s) in degrees Fahrenheit (numeric).}
}
\value{
Temperature value(s) in degrees Celcius (numeric).
}
\description{
Converts temperature values from degrees Fahrenheit to degrees Celcius
}
\examples{
celcius(59)
celcius(14)
celcius(c(0.14, -40, 39.2))
}
\keyword{celcius}
\keyword{fahrenheit}
