% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{fahrenheit}
\alias{fahrenheit}
\title{Celsius to Fahrenheit}
\usage{
fahrenheit(tempC)
}
\arguments{
\item{tempC}{temperature value(s) in degrees Celcius (numeric).}
}
\value{
Temperature value(s) in degrees Fahrenheit (numeric).
}
\description{
Converts temperature values from degrees Celcius to degrees Fahrenheit
}
\examples{
fahrenheit(15)
fahrenheit(-10)
fahrenheit(c(-17.7, -40, 4))
}
\keyword{celcius}
\keyword{fahrenheit}
