% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStationData_single.R
\name{getStationData_single}
\alias{getStationData_single}
\title{Data retrieval from NOAA's Integrated Surface Database (ISD) for single station}
\usage{
getStationData_single(years, usaf, wban)
}
\arguments{
\item{years}{year range (numeric).}

\item{usaf}{station USAF (United States Air Force) ID number (numeric).}

\item{wban}{station WBAN (Weather Bureau Army Navy) ID number (numeric).}
}
\value{
Returns a data frame with the data for the available years in the range given.
}
\description{
Retrieves climate data from a single ISD station for a given year range.
\code{getStationData_single()} expands on rnoaa's \code{isd()} function, by allowing to append data for multiple years in the same data frame and catching and ignoring errors.
}
\note{
USAF and WBAN numbers for a station can be found with the function \code{getStationInfo()} if the name of the station is known, or with rnoaa's \code{isd_stations_search()} if the station location is known.
}
\examples{
stmatthewsIsland  <- getStationData_single(1945, 749231, 99999)
adakIsland        <- getStationData_single(2010:2018, 997380, 99999) # With error message
mtl               <- getStationData_single(1954:1956, 713721, 99999) # With warning message
borabora          <- getStationData_single(1942:1945, 919300, 99999) # With warning and error message
}
