% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isdTempSummary_daily.R
\name{isdTempSummary_daily}
\alias{isdTempSummary_daily}
\title{Daily temperature summaries from NOAA's Integrated Surface Database (ISD)}
\usage{
isdTempSummary_daily(isd_station_data, f = FALSE)
}
\arguments{
\item{isd_station_data}{data frame containing ISD station data, as outputed by rnoaa's \code{isd()} or risd's \code{getStationData_single()} or \code{getStationData_all()}.}

\item{f}{option to convert to Fahrenheit (default is \code{FALSE}).}
}
\value{
Returns a data frame with the daily temperature summaries for the station data.
}
\description{
Computes daily temperature summaries (mean, max and min) from ISD station climate data
}
\examples{
stmattTempSumm <-  isdTempSummary_daily(stmatthewsIsland1940)           # Data retrieved with getStationData_single()
stmattTempSummF <- isdTempSummary_daily(stmatthewsIsland1940, f = TRUE) # Same as above, but in Fahrenheit
icelandTempSumm <- isdTempSummary_daily(icelandSubset)                  # Data retrieved with getStationData_all()
}
